requirements.txt
discord.py[voice]
rich
youtube-dl
aiohttp
bs4
requests
forex_python
mysql.connector